requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.25.1
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
cassandra-driver==3.29.0
cassio==0.1.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect==1.2.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.3
dataclasses-json-speakeasy==0.5.11
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
emoji==2.10.1
entrypoints==0.4
Faker==24.4.0
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
filetype==1.2.0
fonttools==4.50.0
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanize==4.9.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jq==1.6.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.14
langchain-anthropic==0.1.8
langchain-community==0.0.31
langchain-core==0.1.42
langchain-experimental==0.0.51
langchain-google-genai==1.0.1
langchain-mistralai==0.1.2
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langdetect==1.0.9
langsmith==0.1.38
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
more-itertools==10.2.0
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
orjson==3.10.0
packaging==23.2
pandas==2.2.0
pillow==10.3.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
pyparsing==3.1.2
pypdf==4.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2024.1.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.3
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
st-annotated-text==4.0.1
streamlit==1.33.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
transformers==4.39.3
trulens-eval==0.27.2
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.3
unstructured-client==0.17.0
urllib3==2.1.0
validators==0.22.0
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
